# backend/tests/baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "GroupThreadInfo": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "avatar_url"
        ],
        "title": "GroupThreadInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageCreate": {
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "client_message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          }
        },
        "required": [
          "body"
        ],
        "title": "MessageCreate",
        "type": "object"
      },
      "MessageRead": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "client_message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted": {
            "default": false,
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Id"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender_id",
          "body",
          "created_at",
          "read_at"
        ],
        "title": "MessageRead",
        "type": "object"
      },
      "MessageThreadRead": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupThreadInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "latest_message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Id"
          },
          "partner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfilePublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "conversation_id",
          "kind",
          "latest_message"
        ],
        "title": "MessageThreadRead",
        "type": "object"
      },
...
      "ProfilePublic": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "campus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus"
          },
          "class_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Year"
          },
          "country_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country State"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "title": "Interests",
            "type": "array"
          },
          "is_hidden": {
            "title": "Is Hidden",
            "type": "boolean"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "languages_learning": {
            "items": {
              "type": "string"
            },
            "title": "Languages Learning",
            "type": "array"
          },
          "languages_spoken": {
            "items": {
              "type": "string"
            },
            "title": "Languages Spoken",
            "type": "array"
          },
          "looking_for": {
            "items": {
              "type": "string"
            },
            "title": "Looking For",
            "type": "array"
          },
          "looking_for_codes": {
            "items": {
              "type": "string"
            },
            "title": "Looking For Codes",
            "type": "array"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "profile_completed": {
            "title": "Profile Completed",
            "type": "boolean"
          },
          "pronouns": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronouns"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "display_name",
          "pronouns",
          "major",
          "class_year",
          "country_state",
          "campus",
          "bio",
          "avatar_url",
          "is_hidden",
          "is_verified",
          "profile_completed",
          "interests",
          "languages_spoken",
          "languages_learning",
          "looking_for",
          "looking_for_codes"
        ],
        "title": "ProfilePublic",
        "type": "object"
      },
...
      "UnreadSummary": {
        "description": "Total unread + per-conversation counts (conversations with 0 unread are omitted).",
        "properties": {
          "per_conversation": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "title": "Per Conversation",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "per_conversation"
        ],
        "title": "UnreadSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LC Connect API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/messages/threads": {
      "get": {
        "description": "The unified inbox \u2014 DM and group threads, newest activity first.",
        "operationId": "list_threads_api_v1_messages_threads_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageThreadRead"
                  },
                  "title": "Response List Threads Api V1 Messages Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Threads",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/threads/{match_id}": {
      "get": {
        "description": "Newest-first page of a conversation. Pass the oldest row's (created_at, id) as\n`before_*` to fetch the next older page (keyset pagination).",
        "operationId": "get_thread_api_v1_messages_threads__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before_created_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Created At"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageRead"
                  },
                  "title": "Response Get Thread Api V1 Messages Threads  Match Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Thread",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "send_message_api_v1_messages_threads__match_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/threads/{match_id}/sync": {
      "get": {
        "description": "Oldest-first messages after a cursor \u2014 reconnect catch-up.",
        "operationId": "sync_thread_endpoint_api_v1_messages_threads__match_id__sync_get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_created_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "After Created At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "After Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageRead"
                  },
                  "title": "Response Sync Thread Endpoint Api V1 Messages Threads  Match Id  Sync Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sync Thread Endpoint",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/unread-summary": {
      "get": {
        "description": "Total + per-conversation unread counts \u2014 seeds the tab + per-row badges. Keyed by the\nclient-facing addressing id (match id for DMs, conversation id for groups).",
        "operationId": "get_unread_summary_api_v1_messages_unread_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadSummary"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Unread Summary",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/{message_id}": {
      "delete": {
        "description": "Delete a message for everyone (soft delete \u2192 tombstone). Sender anywhere; group admins in\ntheir group. Fans out a `message.deleted` event so open chats update live.",
        "operationId": "delete_message_endpoint_api_v1_messages__message_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Message Endpoint",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}